refactor: move logic modules to folder modules and delte extra imports
