+++ src/App.tsx
@@ -1,4 +1,3 @@
-import Header from 'components/UI/header/Header';
 import React from 'react';
 import { RouterProvider } from 'react-router-dom';
 import router from 'router/router';
