--- src/components/UI/CardsList/CardsList.module.css
@@ -1,13 +1,0 @@
-.CardsList {
-  display: flex;
-  flex-wrap: wrap;
-  row-gap: 30px;
-  margin: 0 -15px;
-}
-
-@media (max-width: 992px) {
-  .CardsList {
-    justify-content: center;
-    margin: 0 -10px;
-  }
-}
