--- src/components/UI/layout/Layout.module.css
@@ -1,4 +1,0 @@
-.main {
-  padding: 30px 0;
-  flex: 1 1 auto;
-}
