+++ src/components/searchBar/SearchBar.module.css
@@ -1,0 +1,36 @@
+.SearchBar {
+  position: relative;
+  margin-bottom: 30px;
+}
+
+.SearchForm {
+  position: relative;
+  display: flex;
+  width: 100%;
+  height: 40px;
+}
+
+.SearchInput {
+  width: 100%;
+  height: 100%;
+  padding: 5px;
+  border: 3px solid #00b4cc;
+  border-right: none;
+  border-radius: 5px 0 0 5px;
+  outline: none;
+  font-size: 16px;
+  color: #000;
+}
+
+.SearchButton {
+  width: 70px;
+  height: 100%;
+  padding: 3px 5px;
+  border: 1px solid #00b4cc;
+  background: #00b4cc;
+  border-radius: 0 5px 5px 0;
+  font-size: 16px;
+  text-align: center;
+  color: #fff;
+  cursor: pointer;
+}
