+++ src/modules/layout/Layout.module.css
@@ -1,0 +1,4 @@
+.main {
+  padding: 30px 0;
+  flex: 1 1 auto;
+}
