+++ src/router/router.tsx
@@ -1,11 +1,9 @@
-import CardsList from 'components/UI/CardsList/CardsList';
-import Header from 'components/UI/header/Header';
-import Layout from 'components/UI/layout/Layout';
+import React from 'react';
+import { createBrowserRouter, Navigate } from 'react-router-dom';
+import Layout from 'modules/layout/Layout';
 import AboutPage from 'pages/AboutPage';
 import MainPage from 'pages/MainPage';
 import NotFoundPage from 'pages/NotFoundPage';
-import React from 'react';
-import { createBrowserRouter, Navigate } from 'react-router-dom';
 
 const router = createBrowserRouter([
   {
